Completed User Story 4.
As a client of a bank,
So that I know what was going on with my account
I would like to be able to print my bank statement, that will show all my transactions, dates of these transactions, and track my balance.

Red / Green. Refactoring WIP. Created Printer class with printing responsibility, and separate Transaction class, responsible for storing one transaction. Currently transactions are stored as an array in Account class instance.